fix: migrate FileUtils to VFS
I replicated the original behavior, for example, writeFile should automatically create dirs + overwrite the file (if exists), readText normalizes line endings + limits the output to 10k chars, etc. All those details should be covered by unit test.

Also, I removed some usages of UriUtils because if we treat "file URIs" as URLs, we can use refreshAndFindFileByUrl. It works, but we should keep that in mind.

### Extra: Manual testing

(I did it myself on top of the unit tests, but I’m posting the instructions here in case someone wants to reproduce it)

* writeFile - run Continue in agentic mode and ask smth like: "create scss/test.scss"
* listDir (+ getFileStats) - ask: "list files in project root"
* openFile/showFile - edit a rule using the toolbar (pen icon)
* fileExists - is invoked on startup multiple times
* readFile - is invoked on startup multiple times (put some files into .continue/*)